Extract modal state management to re-usable component
