+++ projects/js-packages/publicize-components/src/components/connection-management/index.tsx
@@ -1,12 +1,12 @@
 import { Button } from '@automattic/jetpack-components';
 import { Disabled } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
-import { useCallback, useEffect, useState } from '@wordpress/element';
+import { useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import classNames from 'classnames';
 import useSocialMediaConnections from '../../hooks/use-social-media-connections';
 import { store } from '../../social-store';
-import { ManageConnectionsModal } from '../manage-connections-modal';
+import { ManageConnectionsModalWithTrigger as ManageConnectionsModal } from '../manage-connections-modal';
 import { SupportedService, useSupportedServices } from '../services/use-supported-services';
 import { ConnectionInfo } from './connection-info';
 import styles from './style.module.scss';
@@ -14,20 +14,15 @@ import styles from './style.module.scss';
 const ConnectionManagement = ( { className = null } ) => {
 	const { refresh } = useSocialMediaConnections();
 
-	const { connections, deletingConnections, updatingConnections, keyringResult } = useSelect(
-		select => {
-			const { getConnections, getDeletingConnections, getUpdatingConnections, getKeyringResult } =
-				select( store );
+	const { connections, deletingConnections, updatingConnections } = useSelect( select => {
+		const { getConnections, getDeletingConnections, getUpdatingConnections } = select( store );
 
-			return {
-				keyringResult: getKeyringResult(),
-				connections: getConnections(),
-				deletingConnections: getDeletingConnections(),
-				updatingConnections: getUpdatingConnections(),
-			};
-		},
-		[]
-	);
+		return {
+			connections: getConnections(),
+			deletingConnections: getDeletingConnections(),
+			updatingConnections: getUpdatingConnections(),
+		};
+	}, [] );
 
 	connections.sort( ( a, b ) => {
 		if ( a.service_name === b.service_name ) {
@@ -35,8 +30,6 @@ const ConnectionManagement = ( { className = null } ) => {
 		}
 		return a.service_name.localeCompare( b.service_name );
 	} );
-
-	const [ isModalOpen, setIsModalOpen ] = useState( false );
 
 	useEffect( () => {
 		refresh();
@@ -50,20 +43,6 @@ const ConnectionManagement = ( { className = null } ) => {
 		},
 		{}
 	);
-
-	const closeModal = useCallback( () => {
-		setIsModalOpen( false );
-	}, [] );
-	const openModal = useCallback( () => {
-		setIsModalOpen( true );
-	}, [] );
-
-	const shouldModalBeOpen =
-		isModalOpen ||
-		// It's possible that when reconnecting a connection from within the modal,
-		// the user closes the modal immediately, without waiting for the confirmation,
-		// in that case we should show the modal again when the keyringResult is set.
-		keyringResult?.ID;
 
 	return (
 		<div className={ classNames( styles.wrapper, className ) }>
@@ -82,7 +61,6 @@ const ConnectionManagement = ( { className = null } ) => {
 										<ConnectionInfo
 											connection={ connection }
 											service={ servicesByName[ connection.service_name ] }
-											onConfirmReconnect={ openModal }
 										/>
 									</Disabled>
 								</li>
@@ -91,10 +69,13 @@ const ConnectionManagement = ( { className = null } ) => {
 					</ul>
 				</>
 			) : null }
-			<Button onClick={ openModal } variant={ connections.length ? 'secondary' : 'primary' }>
-				{ __( 'Connect an account', 'jetpack' ) }
-			</Button>
-			{ shouldModalBeOpen ? <ManageConnectionsModal onCloseModal={ closeModal } /> : null }
+			<ManageConnectionsModal
+				trigger={
+					<Button variant={ connections.length ? 'secondary' : 'primary' }>
+						{ __( 'Connect an account', 'jetpack' ) }
+					</Button>
+				}
+			/>
 		</div>
 	);
 };
